vmount: fix implicit double promotion

+++ src/modules/vmount/output_mavlink.cpp
@@ -80,8 +80,8 @@ int OutputMavlink::update(const ControlData *control_data)
 		vehicle_command.param3 = 0.0f;
 		vehicle_command.param4 = 0.0f;
 
-		vehicle_command.param5 = static_cast<float>(control_data->type_data.angle.frames[0]);
-		vehicle_command.param6 = static_cast<float>(control_data->type_data.angle.frames[1]);
+		vehicle_command.param5 = static_cast<double>(control_data->type_data.angle.frames[0]);
+		vehicle_command.param6 = static_cast<double>(control_data->type_data.angle.frames[1]);
 		vehicle_command.param7 = static_cast<float>(control_data->type_data.angle.frames[2]);
 
 		_vehicle_command_pub.publish(vehicle_command);
